# medi-care-application/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Hospital Management System
  description: |-
    This helps to manage the hospital
  termsOfService: http://swagger.io/terms/
  contact:
    email: aravind@monkeeys.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: localhost:8080/hospital
...
paths:
...
  /v1/api/payroll:
    post:
      tags:
        - PayRoll
      summary: Creation of new Payroll
      description: Return the list values
      operationId: createPayrollForEmployee
      requestBody:
        description: Payroll creation payload
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayRoll'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayRoll'
        '400':
          description: Invalid Payroll
  /v1/api/payroll/{id}:
    get:
      tags:
        - PayRoll
      summary: Get PayRoll by id
      description: Appoinment details
      operationId: getPayRollById
      parameters:
        - name: id
          in: path
          description: PayRoll Id
          required: true
          explode: true
          schema:
            type: string
            default: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayRoll'
        '400':
          description: Invalid Appointment id
    put:
      tags:
        - PayRoll
      summary: Update PayRoll
      description: Update PayRoll data
      operationId: updatePayRoll
      requestBody:
        description: Update an existent PayRoll
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayRoll'
        required: true
      parameters:
        - name: id
          in: path
          description: PayRoll Id
          required: true
          explode: true
          schema:
            type: string
            default: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayRoll'
        '400':
          description: Invalid Appointment id
    delete:
      tags:
        - PayRoll
      summary: Delete PayRoll
      description: Delete PayRoll by id
      operationId: deletePayRollById
      parameters:
        - name: id
          in: path
          description: PayRoll Id
          required: true
          explode: true
          schema:
            type: string
            default: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: string
                example: "PayRoll removed successfully"
        '400':
          description: Invalid PayRoll id
  /v1/api/payroll-dash-summary:
    get:
      tags:
        - PayRoll
      summary: Get PayRoll by id
      description: Appoinment details
      operationId: getPayRollSummaryForDash
      parameters:
        - name: id
          in: query
          description: Hospital Id
          required: true
          explode: true
          schema:
            type: string
            default: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayRollSummary'
        '400':
          description: Invalid Appointment id
  /v1/api/payroll-leave-summary:
    get:
      tags:
        - PayRoll
      summary: Get Leave summary for all employees
      description: Leave summary details
      operationId: getPayRollLeaveSummaryDetails
      parameters:
        - name: id
          in: query
          description: Hospital Id
          required: true
          explode: true
          schema:
            type: string
            default: available
        - name: month
          in: query
          description: Month
          required: true
          explode: true
          schema:
            type: number
            default: available
        - name: year
          in: query
          description: Year
          required: true
          explode: true
          schema:
            type: number
            default: available
        - name: from
          in: query
          description: From Date
          required: true
          explode: true
          schema:
            type: string
            default: available
        - name: to
          in: query
          description: To Date
          required: true
          explode: true
          schema:
            type: string
            default: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayRollSummary'
        '400':
          description: Invalid Appointment id
  /v1/api/payroll/by-employee/{id}:
    get:
      tags:
        - PayRoll
      summary: Get PayRoll by employee id
      description: Appoinment details
      operationId: getPayRollByEmployeeId
      parameters:
        - name: id
          in: path
          description: Employee Id
          required: true
          explode: true
          schema:
            type: string
            default: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayRoll'
        '400':
          description: Invalid employee id
  /v1/api/payroll/generate:
    post:
      tags:
        - PayRoll
      summary: Creation of new Payroll History
      description: Return the list values
      operationId: createPayRollGenerationHistory
      requestBody:
        description: Payroll creation payload
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayRollGenerationHistory'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayRollGenerationHistory'
        '400':
          description: Invalid Payroll
    get:
      tags:
        - PayRoll
      summary: Payroll generation
      description: Payroll generation
      operationId: generatePayRollForEmployees
      parameters:
        - name: month
          in: query
          description: Payroll month
          required: true
          explode: true
          schema:
            type: string
            default: available
        - name: year
          in: query
          description: Payroll year
          required: false
          explode: true
          schema:
            type: string
            default: available
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PayRollGenerationHistory'
        '400':
          description: Invalid request
...
components:
  schemas:
...
    UserBasiInfo:
      type: object
      properties:
        user_id:
          type: string
          example: fb17eecf-5ec6-4c13-a5bd-5ce1451472d0
        first_name:
          type: string
          example: Sastika
        last_name:
          type: string
          example: Aravind
        role:
          $ref: '#/components/schemas/Role'
        gender:
          type: string
          example: Male
          enum:
            - Male
            - FeMale
            - Others
        employment_details:
          $ref: '#/components/schemas/EmploymentDetails'
...
    Role:
      type: object
      properties:
        id:
          type: string
          example: 4acd0b30-0ac1-4d0f-938b-c7b86d804375
        name:
          type: string
          example: ROLE_USER
        description:
          type: string
          example: "User role"
    EmploymentDetails:
      type: object
      properties:
        id:
          type: string
          example: 4acd0b30-0ac1-4d0f-938b-c7b86d804375
        joining_date:
          type: string
          format: date-time
          pattern: "^(\\d{4})-(\\d{2})-(\\d{2})T((\\d{2}):(\\d{2}):(\\d{2}))\\.(\\d{3})Z$"
          description: "User joining date."
          example: "2024-11-05T00:00:00Z"
        relieving_date:
          type: string
          format: date-time
          pattern: "^(\\d{4})-(\\d{2})-(\\d{2})T((\\d{2}):(\\d{2}):(\\d{2}))\\.(\\d{3})Z$"
          description: "User joining date."
          example: "2024-11-05T00:00:00Z"
        department:
          type: string
          example: "CLINIC_STAFF"
        salary:
          type: number
          example: 18500
        payroll:
          $ref: '#/components/schemas/PayRoll'
    PayRoll:
      type: object
      properties:
        id:
          type: string
          example: 4acd0b30-0ac1-4d0f-938b-c7b86d804375
        user:
          $ref: '#/components/schemas/UserBasiInfo'
        payment_date:
          type: string
          format: date-time
          pattern: "^(\\d{4})-(\\d{2})-(\\d{2})T((\\d{2}):(\\d{2}):(\\d{2}))\\.(\\d{3})Z$"
          description: "Appointment Booking Date"
          example: "2024-11-05T00:00:00Z"
        gross_amount:
          type: number
          format: float
        deductions:
          type: number
          format: float
        net_amount:
          type: number
          format: float
    PayRollSummary:
      type: object
      properties:
        total_amount:
          type: number
          example: 10
        salary_per_month:
          type: number
          format: float
          example: 12563.58
        gradWise:
          type: array
          items:
            $ref: '#/components/schemas/GradeWiseSummary'
    PayRollGenerationHistory:
      type: object
      properties:
        emp_id:
          type: string
          example: 4acd0b30-0ac1-4d0f-938b-c7b86d804375
        emp_name:
          type: string
          example: aravind
        leave:
          type: integer
          example: 10
        cross_salary:
          type: number
          example: 15000.63
        deductions:
          type: number
          example: 1596.32
        deduction_summary:
          type: array
          items:
            $ref: '#/components/schemas/DeductionSummary'
        net_salary:
          type: number
          example: 14899.90
        status:
          type: string
          example: APPROVED
          enum:
            - APPROVED
            - REJECTED
    GradeWiseSummary:
      type: object
      properties:
        role:
          type: string
          example: role_name
        count:
          type: number
          example: 10
...
    DeductionSummary:
      type: object
      properties:
        reason:
          type: string
          example: TDS
        amount:
          type: number
          example: 1800.30
...